Add utility to detect container runtime
A new utility script has been added to automatically detect the container runtime. Now, instead of hard coding the use of podman or docker, the script will check which container runtime is available and correspondingly set the 'CONTAINER_RUNTIME' environment variable. This simplifies and streamlines the process of running and managing containers.